Implement Server's Message nonce increment (WIP)
Start debugging this new step in the chain.

This seems to have broken both the Client's Hello (I see a stack trace
about this in the logs) and the Server's Cookie building (I see
complaints about it being missing, also in the Client logs.

And the client's sending multiple Hello Packets.

So I've managed to botch up something.