Fixed a couple of bugs
I missed a reference to shared.logging when I refactored that away.

There was a closing parenthesis in the wrong spot which wound up
setting the client-extension-load-time to nil.

The former kept it from publishing to clojars.

The latter seemed to wreak all kinds of havoc with the Cookie packet
coming back from the server. Well, the problems seemed to go away
after I fixed it.