Converted crypto/open-box to return byte-array
This has been bothering me for quite a while now. It came to a head
while I was cleaning up a bunch of reflection warnings.

I don't think those are all cleared up (yagni still reports problems,
but it's silent about their nature).

Fought with a few unit tests. Gave up on several as "not worth the
trouble."