Attempt to clean up versioning
I want to be able to tag a real release, then move forward with it
based on hash tags.

This flies in the face of semver, but semver is pretty broken
anyway. As long as I have a commit hash, I might as well use it. Even
though it's meaningless.

This kind of approach probably makes sense for something like
clojurescript that's trying to keep its builds in lockstep with
clojure's. It doesn't seem to make as much sense for a project like
this.

This needs hammock time.

Making it work, verifying it, and then thinking it through seems like
a terrible idea, but that's how this is going down.