Fix up server's pre-state-options spec (WIP)
Well, possibly.

Have a unit test that tries to use that spec and fails with an error
that doesn't make much sense to me. This was an attempt to fix that,
but it might have been a no-op.

It really depends on how the s/def macro is implemented.

Incremented the test.check dependency version.

Along with the default generated version for running inside docker.

Updated the README that this isn't *quite* vaporware.